src/app/lunr/filters/stopword.ts
/*
 * Copyright (c) 2022 dataplay.io, all rights reserved.
 */

import { IndexPipelineFunction } from '../index-pipeline';

/**
 * defaultStopWordFilter is an English language stop words filter, any words
 * contained in the stop word list will not be passed through the filter.
 *
 * This is intended to be used in the Pipeline. If the token does not pass the
 * filter then undefined will be returned.
 * Currently this StopwordFilter using dictionary to do O(1) time complexity stop word filtering.
 */
export function defaultStopWordFilter(): IndexPipelineFunction {
  return stopWordFilter(...DefaultStopWords);
}

/**
 * Construct a stopword filter provided a list of stopwords. You may generate your own,
 * or use the provided DefaultStopWords to seed your own filter lambda.
 */
export function stopWordFilter(...stopWords: string[]): IndexPipelineFunction {
  const stopWordMap: Set<string> = new Set<string>(stopWords);
  return (w) => !w || stopWordMap.has(w) ? undefined : w;
}

/**
 * The default list of stop words.
 */
export const DefaultStopWords: string[] = [
  '',
  'a',
  'able',
  'about',
  'across',
  'after',
  'all',
  'almost',
  'also',
  'am',
  'among',
  'an',
  'and',
  'any',
  'are',
  'as',
  'at',
  'be',
  'because',
  'been',
  'but',
  'by',
  'can',
  'cannot',
  'could',
  'dear',
  'did',
  'do',
  'does',
  'either',
  'else',
  'ever',
  'every',
  'for',
  'from',
  'get',
  'got',
  'had',
  'has',
  'have',
  'he',
  'her',
  'hers',
  'him',
  'his',
  'how',
  'however',
  'i',
  'if',
  'in',
  'into',
  'is',
  'it',
  'its',
  'just',
  'least',
  'let',
  'like',
  'likely',
  'may',
  'me',
  'might',
  'most',
  'must',
  'my',
  'neither',
  'no',
  'nor',
  'not',
  'of',
  'off',
  'often',
  'on',
  'only',
  'or',
  'other',
  'our',
  'own',
  'rather',
  'said',
  'say',
  'says',
  'she',
  'should',
  'since',
  'so',
  'some',
  'than',
  'that',
  'the',
  'their',
  'them',
  'then',
  'there',
  'these',
  'they',
  'this',
  'tis',
  'to',
  'too',
  'twas',
  'us',
  'wants',
  'was',
  'we',
  'were',
  'what',
  'when',
  'where',
  'which',
  'while',
  'who',
  'whom',
  'why',
  'will',
  'with',
  'would',
  'yet',
  'you',
  'your'
];
